requirements.txt
altair
pandas
streamlit
ai21
python-dotenv
asyncio
httpx-oauth
streamlit-cookies-manager
streamlit-extras
